MODELINKS-80 Add patch and get linking rule endpoints (#59)
- Migrate linking rule id to integer instead of bigint
- Add autoLinkingEnabled field to linking rule
- Set default autoLinkingEnabled values for all default rules

+++ src/test/java/org/folio/support/TestDataUtils.java
@@ -273,7 +273,7 @@ public class TestDataUtils {
           .naturalId(naturalId)
           .build())
         .linkingRule(InstanceAuthorityLinkingRule.builder()
-          .id((long) TAGS_TO_RULE_IDS.get(tag))
+          .id(TAGS_TO_RULE_IDS.get(tag))
           .bibField(tag)
           .build())
         .build();
